Moved client class to another file
